Refactored account activity
