+++ app/src/main/AndroidManifest.xml
@@ -69,7 +69,7 @@
             android:theme="@style/AppTheme.NoActionBar" />
 
         <activity
-            android:name=".activities.AccountActivity"
+            android:name=".views.account.AccountView"
             android:theme="@style/AppTheme.NoActionBar" />
 
         <activity
@@ -77,7 +77,7 @@
             android:theme="@style/AppTheme.NoActionBar" />
 
         <activity
-            android:name=".activities.AboutActivity"
+            android:name=".views.about.AboutView"
             android:theme="@style/AppTheme.NoActionBar" />
 
         <activity
